renamed mini app report form
